Add `CXCallable` protocol for handling duplicated `CXKit` methods in both `VoiceCallVC` and `VideoCallVC`
