+++ QuickStart/View/Base.lproj/Main.storyboard
@@ -69,7 +69,7 @@
                                 <rect key="frame" x="185" y="361" width="44" height="21"/>
                                 <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                 <constraints>
-                                    <constraint firstAttribute="width" constant="44" id="p5Y-On-zHU"/>
+                                    <constraint firstAttribute="width" relation="greaterThanOrEqual" constant="44" id="p5Y-On-zHU"/>
                                 </constraints>
                                 <fontDescription key="fontDescription" type="system" pointSize="17"/>
                                 <color key="textColor" red="0.0" green="0.0" blue="0.0" alpha="0.5" colorSpace="custom" customColorSpace="calibratedRGB"/>
